Remove format on save
this was making me pull my hair out for about the past week. why is this
enabled by default, and why is it only disabled in c/cpp files (ok maybe
        i did that)? but still. if you like formatting on save i
legitimately dont believe you are a human being. you are an android
mimicing the actions of a human. you are not biological in nature.